New build and wheel file with corrected K41Cs dict
