+++ Example Scripts/Output_multithreaded_DC.py
