+++ diatom.egg-info/requires.txt
@@ -1,0 +1,3 @@
+numpy>=1.15
+sympy>=1.4
+scipy>=1.1
